committing

+++ 0x00-hello_world/6-size.c
@@ -14,10 +14,10 @@ int main(void)
 	long int l;
 	long long int l2;
 
-	printf("Size of an int: %lu byte(s)\n", sizeof(i));
-	printf("Size of a char: %lu byte(s)\n", sizeof(c));
-	printf("Size of a float: %lu byte(s)\n", sizeof(f));
-	printf("Size of a long int: %lu byte(s)\n", sizeof(l));
-	printf("Size of a long long int: %lu byte(s)\n", sizeof(l2));
+	printf("Size of an int: %lu byte(s)\n", (unsigned long)sizeof(i));
+	printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(c));
+	printf("Size of a float: %lu byte(s)\n", (unsigned long)sizeof(f));
+	printf("Size of a long int: %lu byte(s)\n", (unsigned long)sizeof(l));
+	printf("Size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(l2));
 	return (0);
 }
